Created helper function asset folder
Help functions do three things:
1. Combines all purchases with same category
2. Attaches a color for each category for chart purposes
3. Attaches each category's budget to the array
