+++ client/components/ChartComponents/BarGraphCategory.js
@@ -4,7 +4,7 @@ import { BarChart, Bar, Cell, XAxis, YAxis, CartesianGrid, Tooltip, Legend, Resp
 import CategoryLegend from './BarGraphCategoryLegend';
 import CustomToolTip from './BarGraphCategoryToolTip';
 import barColors from './assets/categoryColors'
-import categoryMerger from './assets/categoryMerger';
+import {budgetMerger, categoryMerger, colorMerger} from './assets/mergerHelperFuncs';
 
 const BarGraphCategory = (props) => {
 
@@ -20,35 +20,8 @@ const BarGraphCategory = (props) => {
   }) || [];
 
   let combinedDataArray = categoryMerger(dataArray) || [];
-
-  combinedDataArray.forEach((category, index) => {
-    //random number from 2000 to 5000
-    category.budget = Math.floor(Math.random() * (5000 - 2000 + 1)) + 2000;
-    category.color = barColors[index];
-  })
-  console.log(categories);
-  console.log(combinedDataArray);
-
-  const budgetForCategory = (arrayFromState, combinedCategories) => {
-    for (let i = 0; i < arrayFromState.length; i++){
-      let categoryFromState = arrayFromState[i];
-      let stateCategoryName = categoryFromState.categoryName; //bank fees, etc.
-
-      for (let j = 0; j < combinedCategories.length; j++){
-        let combinedCategory = combinedCategories[j].category; //bank fees, etc.
-        if (stateCategoryName === combinedCategory){
-          if(categoryFromState.budgetCategories.length !== 0){
-            combinedCategories[j].budget = categoryFromState.budgetCategories[0].budgetForCategory
-          } else {
-            combinedCategories[j].budget = 0;
-          }
-        }
-      }
-    }
-    return combinedCategories;
-  }
-
-  budgetForCategory(categories, combinedDataArray);
+  colorMerger(combinedDataArray);
+  budgetMerger(categories, combinedDataArray);
 
   return (
     <ResponsiveContainer width="100%" height="100%">
